fix(llc): normalize address / xpub before computing A4 account version
