+++ .changeset/giant-eels-hang.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/live-common": patch
+---
+
+fix(llc): normalize hex-encoded before computing A4 account versions
