+++ libs/ledger-live-common/src/bridge/generic-coin-framework/a4/client/accountId.test.ts
@@ -1,45 +1,4 @@
-import { deriveA4AccountId, normalizeAccountKey } from "./accountId";
-
-describe("normalizeAccountKey", () => {
-  it("lowercases EVM 0x address", () => {
-    expect(normalizeAccountKey("0x742d35Cc6634C0532925a3b844Bc454e4438f44e")).toEqual(
-      "0x742d35cc6634c0532925a3b844bc454e4438f44e",
-    );
-  });
-
-  it.each<[string, string, string]>([
-    [
-      "bitcoin",
-      "xpub6CUGRUonZSQ4TWtTMmzXdrXDtypWKiKrhko4egpiMZbpiaQL2jkwSB1icqYh2cfDfVxdx4df189oLKnC5fSwqPfgyP3hooxujYzAu3fDVmz",
-      "xpub6CUGRUonZSQ4TWtTMmzXdrXDtypWKiKrhko4egpiMZbpiaQL2jkwSB1icqYh2cfDfVxdx4df189oLKnC5fSwqPfgyP3hooxujYzAu3fDVmz",
-    ],
-    [
-      "canton",
-      "BobCorp::12208fceed3e5e2ae59cb00f00d7b3a0cc7f1f1c53d44c5c22cd7424f1b79b4d66d::1",
-      "BobCorp::12208fceed3e5e2ae59cb00f00d7b3a0cc7f1f1c53d44c5c22cd7424f1b79b4d66d::1",
-    ],
-    [
-      "cardano",
-      "addr1qx2fxv2umyhttkxyxp8x0dlpdt3k6cwng5pxj3jhsydzer3n0d3vllmyqwsx5wktcd8cc3sq835lu7drv2xwl2wywfgs68faae",
-      "addr1qx2fxv2umyhttkxyxp8x0dlpdt3k6cwng5pxj3jhsydzer3n0d3vllmyqwsx5wktcd8cc3sq835lu7drv2xwl2wywfgs68faae",
-    ],
-    ["ripple", "rHb9CJAWyB4rj91VRWn96DkukG4bwdtyTh", "rHb9CJAWyB4rj91VRWn96DkukG4bwdtyTh"],
-    [
-      "solana",
-      "EkmJ7PEGdPhUH1MSHcjLnz5MN8aJMGrxvfDz5kJKsmYT",
-      "EkmJ7PEGdPhUH1MSHcjLnz5MN8aJMGrxvfDz5kJKsmYT",
-    ],
-    [
-      "stellar",
-      "GBOVKZBEM2YYLOCDCUXJ4IMRKHN4LCJAE7WEAEA2KF562XFAGDBOB64V",
-      "GBOVKZBEM2YYLOCDCUXJ4IMRKHN4LCJAE7WEAEA2KF562XFAGDBOB64V",
-    ],
-    ["tezos", "tz1VSUr8wwNhLAzempoch5d6hLRiTh8Cjcjb", "tz1VSUr8wwNhLAzempoch5d6hLRiTh8Cjcjb"],
-    ["tron", "TJmV3QHMiDHpFSs9v5BKxJaVxpAV3q6QnF", "TJmV3QHMiDHpFSs9v5BKxJaVxpAV3q6QnF"],
-  ])("keeps %s address verbatim", (_, input, expected) => {
-    expect(normalizeAccountKey(input)).toEqual(expected);
-  });
-});
+import { deriveA4AccountId } from "./accountId";
 
 describe("deriveA4AccountId", () => {
   it("returns known digest for ethereum address (EVM, lowercased before hashing)", () => {
