+++ libs/ledger-live-common/src/bridge/generic-coin-framework/a4/client/accountId.ts
@@ -1,9 +1,6 @@
 import { sha256 } from "@noble/hashes/sha2";
 import { bytesToHex } from "@noble/hashes/utils";
-
-export function normalizeAccountKey(xpubOrAddress: string): string {
-  return /^0x[0-9a-fA-F]+$/.test(xpubOrAddress) ? xpubOrAddress.toLowerCase() : xpubOrAddress;
-}
+import { normalizeAccountKey } from "./utils";
 
 export function deriveA4AccountId(xpubOrAddress: string): string {
   const textEncoder = new TextEncoder();
