+++ libs/ledger-live-common/src/bridge/generic-coin-framework/a4/client/accountVersion.ts
@@ -1,7 +1,8 @@
 import { sha256 } from "@noble/hashes/sha2";
 import { bytesToHex } from "@noble/hashes/utils";
+import { normalizeAccountKey } from "./utils";
 
 export function computeA4AccountVersion(addresses: string[]): string {
-  const payload = [...addresses].sort().join("|");
+  const payload = addresses.map(normalizeAccountKey).sort().join("|");
   return bytesToHex(sha256(new TextEncoder().encode(payload)));
 }
